Migrate to PlatformIO
- Convert to PlatformIO project structure
- Rename server.h/cpp to os_server.h/cpp to fix naming clash
