+++ .gitignore
@@ -5,3 +5,9 @@ wtopts.txt
 *.sh
 testmode.h
 build-1284/*
+
+.pio
+.vscode/.browse.c_cpp.db*
+.vscode/c_cpp_properties.json
+.vscode/launch.json
+.vscode/ipch
