+++ src/os_server.h
@@ -21,8 +21,8 @@
  * <http://www.gnu.org/licenses/>. 
  */
  
-#ifndef _SERVER_H
-#define _SERVER_H
+#ifndef _OS_SERVER_H
+#define _OS_SERVER_H
 
 #if !defined(ARDUINO)
 #include <stdarg.h>
@@ -86,4 +86,4 @@ public:
 };
 
 
-#endif // _SERVER_H
+#endif // _OS_SERVER_H
