--- updater.sh
@@ -1,5 +1,0 @@
-#! /bin/bash
-
-git pull
-./build.sh -s ospi
-/etc/init.d/OpenSprinkler.sh restart
